meter, Win32, fileaccess per block/per frame
Added speed measurements of every block or frame(n blocks) in file read
for win32. Read times for every frame are saved in an array and printed
(average values, variance) in the terminal. The transmission of data
back to CtrlApp has not changed. Note: speed per frame values need to be
checked.